+++ balrog/__init__.py
@@ -1,6 +1,6 @@
 """Balrog public API."""
 
-__version__ = '1.0.1'
+__version__ = '1.1.0'
 
 try:
     from balrog.policy import Policy