+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 1.0.1
 -----
 
-- Pass the context to the get_identity(*args, **kwargs) (olegpidsadnyi)
+- Pass the context to the get_identity (olegpidsadnyi)
 
 1.0.0
 -----