+++ CHANGES.rst
@@ -1,8 +1,8 @@
 Changelog
 =========
 
-dev
----
+1.1.0
+-----
 
 - Policy will keep track of all the permissions and raise PermissionNotFound when called with unknown permission (hvdklauw)